% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_fsvsim.R
\name{corelement}
\alias{corelement}
\title{Extract "true" model-implied correlations of two series only}
\usage{
corelement(x, i, j, these = seq_len(nrow(x$y)))
}
\arguments{
\item{x}{Object of class \code{'fsvsim'}, usually resulting from a call
of the function \code{\link{fsvsim}}.}

\item{i}{Index of component series 1.}

\item{j}{Index of component series 2.}

\item{these}{Vector indicating which points in time should be extracted.}
}
\value{
Vector with the requested correlations.
}
\description{
\code{corelement} extracts the model-implied (time-varying) correlations between
(exactly) two component series.
}
\seealso{
Other simulation: 
\code{\link{cormat.fsvsim}()},
\code{\link{covelement}()},
\code{\link{covmat.fsvsim}()}
}
\concept{simulation}
