% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_fsvsim.R
\name{covelement}
\alias{covelement}
\title{Extract "true" model-implied covariances of two series only}
\usage{
covelement(x, i, j, these = seq_len(nrow(x$y)))
}
\arguments{
\item{x}{Object of class \code{'fsvsim'}, usually resulting from a call
of the function \code{\link{fsvsim}}.}

\item{i}{Index of component series 1.}

\item{j}{Index of component series 2.}

\item{these}{Vector indicating which points in time should be extracted,
defaults to all.}
}
\value{
Vector with the requested covariances.
}
\description{
\code{covelement} extracts the model-implied (time-varying) covariances between
(exactly) two component series.
}
\seealso{
Other simulation: 
\code{\link{corelement}()},
\code{\link{cormat.fsvsim}()},
\code{\link{covmat.fsvsim}()}
}
\concept{simulation}
