% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{covmat}
\alias{covmat}
\title{Generic extraction of covariance matrix}
\usage{
covmat(x, ...)
}
\arguments{
\item{x}{An object of class \code{fsvdraws} or \code{fsvsim}.}

\item{...}{Arguments to be passed to methods.}
}
\value{
Structure containing the model-implied covariance matrix.
}
\description{
Generic function for extracting model-implied covariance matrices, either
from the MCMC output, or from the simulated model. Details about the
function's behavior can be found in  \code{\link{covmat.fsvdraws}}
(the function invoked when applied to MCMC output) or
\code{\link{covmat.fsvsim}} (the function invoked when applied to a
simulated model.
}
\seealso{
Other generics: 
\code{\link{cormat}()}
}
\concept{generics}
