% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_other.R
\name{expweightcov}
\alias{expweightcov}
\title{Computes the empirical exponentially weighted covariance matrix}
\usage{
expweightcov(dat, alpha = 4/126, hist = 180)
}
\arguments{
\item{dat}{Matrix containing the data, with \code{n} rows
(points in time) and \code{m} columns (component series).}

\item{alpha}{Speed of decay.}

\item{hist}{How far to go back in time?}
}
\value{
A \code{m} times \code{m} covariance matrix estimate.
}
\description{
A common way to get estimates for time-varying covariance matrices
is the compute the exponentially weighted empirical covariance matrix.
}
