% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{facloadcredplot}
\alias{facloadcredplot}
\title{Displays bivariate marginal posterior distribution of factor loadings.}
\usage{
facloadcredplot(x, quants = c(0.01, 0.99))
}
\arguments{
\item{x}{Object of class \code{'fsvdraws'}, usually resulting from a call
to \code{\link{fsvsample}}.}

\item{quants}{Posterior quantiles to be plotted.}
}
\value{
Returns \code{x} invisibly.
}
\description{
\code{facloadcredplot} illustrates the bivariate marginals of the 
factor loadings distribution. It is a monochrome variant of
\code{\link{facloadpairplot}}.
}
\seealso{
Other plotting: 
\code{\link{comtimeplot}()},
\code{\link{corimageplot}()},
\code{\link{corplot}()},
\code{\link{cortimeplot}()},
\code{\link{evdiag}()},
\code{\link{facloaddensplot}()},
\code{\link{facloadpairplot}()},
\code{\link{facloadpointplot}()},
\code{\link{facloadtraceplot}()},
\code{\link{logvartimeplot}()},
\code{\link{paratraceplot}()},
\code{\link{plot.fsvdraws}()},
\code{\link{plotalot}()},
\code{\link{voltimeplot}()}
}
\concept{plotting}
