% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_fsvdraws.R
\name{fsvdsavs}
\alias{fsvdsavs}
\title{DSAVS}
\usage{
fsvdsavs(fsvdraws_in)
}
\arguments{
\item{fsvdraws_in}{Object of class \code{'fsvdraws'}, usually resulting from a call
to \code{\link{fsvsample}}.}
}
\description{
DSAVS
}
