% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_fsvdraws.R
\name{fsvdsavs2}
\alias{fsvdsavs2}
\title{DSAVS2}
\usage{
fsvdsavs2(fsvdraws, each, store_everything)
}
\arguments{
\item{fsvdraws_in}{Object of class \code{'fsvdraws'}, usually resulting from a call
to \code{\link{fsvsample}}.}
}
\description{
DSAVS2
}
