% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_fsvdraws.R
\name{fsvdsavs3}
\alias{fsvdsavs3}
\title{DSAVS3}
\usage{
fsvdsavs3(fsvdraws, each, store_everything, type, nu = 2)
}
\arguments{
\item{fsvdraws_in}{Object of class \code{'fsvdraws'}, usually resulting from a call
to \code{\link{fsvsample}}.}
}
\description{
DSAVS3
}
