% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_other.R
\name{ledermann}
\alias{ledermann}
\title{Ledermann bound for the number of factors}
\usage{
ledermann(m)
}
\arguments{
\item{m}{Number of component series.}
}
\value{
The Ledermann bound, a nonnegative integer.
}
\description{
In the static factor case, the Ledermann bound is the largest
integer rank for which a unique decomposition of the covariance
matrix is possible. (This is the largest possible number of
factors which can be used for \code{\link[stats]{factanal}}.
}
\seealso{
preorder
}
