% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_fsvdraws.R
\name{orderident}
\alias{orderident}
\title{A posteriori factor order identification}
\usage{
orderident(x, method = "averagemeanabs")
}
\arguments{
\item{x}{Object of class \code{'fsvdraws'}, usually resulting from a call
to \code{\link{fsvsample}}.}

\item{method}{Currently only "averagemeanabs" is supported.
This mean that the factors are ordered according to their average
mean absolute loadings size.}
}
\value{
Returns an object of class \code{'fsvdraws'} with adjusted
ordering.
}
\description{
\code{orderident} provides a (very ad-hoc) method for identifying
the ordering of the factors after running the (unrestricted) MCMC
sampler by 
ordering according to the argument \code{method}. Not recommended for
production code (often unstable).
}
\seealso{
Other postprocessing: \code{\link{signident}}
}

