% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.fsvdraws}
\alias{plot.fsvdraws}
\title{Default factor SV plot}
\usage{
\method{plot}{fsvdraws}(x, quantiles = c(0.05, 0.5, 0.95), col = NULL, fsvsimobj = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{'fsvdraws'}, usually resulting from a call
to \code{\link{fsvsample}}.}

\item{quantiles}{Posterior quantiles to be visualized. Must be of length 1 or 3.}

\item{col}{Optional color palette.}

\item{fsvsimobj}{To indicate data generating values in case of simulated
data, pass an optional object of type \code{fsvsim} (usually the result of a
call to \code{\link{fsvsim}}).}

\item{...}{Other arguments will be passed on to \link[corrplot]{corrplot}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Displays the correlation matrix at the last sampling point in time.
}
\seealso{
Other plotting: 
\code{\link{comtimeplot}()},
\code{\link{corimageplot}()},
\code{\link{corplot}()},
\code{\link{cortimeplot}()},
\code{\link{evdiag}()},
\code{\link{facloadcredplot}()},
\code{\link{facloaddensplot}()},
\code{\link{facloadpairplot}()},
\code{\link{facloadpointplot}()},
\code{\link{facloadtraceplot}()},
\code{\link{logvartimeplot}()},
\code{\link{paratraceplot}()},
\code{\link{plotalot}()},
\code{\link{voltimeplot}()}
}
\concept{plotting}
