% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotalot}
\alias{plotalot}
\title{Several factor SV plots useful for model diagnostics}
\usage{
plotalot(x, fsvsimobj = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{'fsvdraws'}, usually resulting from a call
to \code{\link{fsvsample}}.}

\item{fsvsimobj}{To indicate data generating values in case of simulated
data, pass an object of type \code{fsvsim} (usually the result of a
call to \code{\link{fsvsim}}).}

\item{...}{Other arguments will be passed on to the subfunctions.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Draws a collection of plots to explore the posterior distribution
of a fitted factor SV model.
}
\seealso{
Other plotting: 
\code{\link{comtimeplot}()},
\code{\link{corimageplot}()},
\code{\link{corplot}()},
\code{\link{cortimeplot}()},
\code{\link{evdiag}()},
\code{\link{facloadcredplot}()},
\code{\link{facloaddensplot}()},
\code{\link{facloadpairplot}()},
\code{\link{facloadpointplot}()},
\code{\link{facloadtraceplot}()},
\code{\link{logvartimeplot}()},
\code{\link{paratraceplot}()},
\code{\link{plot.fsvdraws}()},
\code{\link{voltimeplot}()}
}
\concept{plotting}
