% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{predsavs_beta}
\alias{predsavs_beta}
\title{Predict bla bla}
\usage{
predsavs_beta(fsvdraws, ahead, penalty_type, nu = 2, type, seed = NULL)
}
\arguments{
\item{fsvdraws}{}

\item{ahead}{}
}
\description{
Predict bla bla
}
