% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_fsvdraws.R
\name{print.fsvdraws}
\alias{print.fsvdraws}
\title{Pretty printing of an fsvsdraws object}
\usage{
\method{print}{fsvdraws}(x, ...)
}
\arguments{
\item{x}{Object of class \code{'fsvdraws'}, usually resulting from a call
of \code{\link{fsvsample}}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Pretty printing of an fsvsdraws object
}
\concept{printing}
